Add HomePageEvent.

+++ lib/app/pages/home/onboarding/onboarding_page_view_state.dart
@@ -5,7 +5,7 @@ class OnboardingPageViewState {
   static const double textLeft = 20;
   static const double textRight = 20;
   static const double textBottom = 45;
-  static const double pageCount = 6;
+  static const int pageCount = 6;
 
   OnboardingPageViewState({required this.language});
 }
